requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
xgboost==1.7.6
tensorflow==2.13.0
keras==2.13.1
ta==0.10.2
python-dotenv==1.0.0
sqlalchemy==2.0.19
matplotlib==3.7.2
plotly==5.17.0
ccxt==4.5.28
schedule==1.2.2
